requirements.txt
black==23.3.0
fastapi==0.95.2
isort==5.12.0
Jinja2==3.1.2
pytest-playwright==0.3.3
uvicorn==0.22.0